refactor(test): move RunControl tests to lifecycle-contract
