+++ src/lifecycle.test.ts
@@ -167,26 +167,6 @@ describe("scheduleMemoryCommit", () => {
   });
 });
 
-describe("createRunControl", () => {
-  test("defaults to no-op callbacks", () => {
-    const rc = createRunControl();
-    expect(rc.shouldYield()).toBe(false);
-    expect(rc.isCancelled()).toBe(false);
-  });
-
-  test("accepts partial overrides", () => {
-    const rc = createRunControl({ shouldYield: () => true });
-    expect(rc.shouldYield()).toBe(true);
-    expect(rc.isCancelled()).toBe(false);
-  });
-
-  test("accepts full overrides", () => {
-    const rc = createRunControl({ shouldYield: () => true, isCancelled: () => true });
-    expect(rc.shouldYield()).toBe(true);
-    expect(rc.isCancelled()).toBe(true);
-  });
-});
-
 describe("runLifecycle yield", () => {
   const baseInput = {
     request: { model: "gpt-5-mini" as const, message: "test", history: [] as never[], useMemory: false },
